+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- New sort option _relevance. At the moment this feature is considered experimental.
+
 ## [0.30.2] - 2021-06-27
 
 ### Fixed